requirements.txt
pymongo
dnspython
requests
python-telegram-bot
telethon